TextToSQL/requirements.txt
streamlit
python-dotenv

langchain
mysql-connector-python
pymysql
langchain_google_genai
langchain_community
